+++ Source/Properties/AssemblyInfo.cs
@@ -27,6 +27,6 @@ using System.Runtime.InteropServices;
 //
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
-[assembly: AssemblyVersion ("1.27.*")]
-[assembly: AssemblyInformationalVersion ("1.27 for KSP v1")]
-[assembly: KSPAssembly ("KIS", 1, 27)]
+[assembly: AssemblyVersion ("1.28.*")]
+[assembly: AssemblyInformationalVersion ("1.28 for KSP v1.11+")]
+[assembly: KSPAssembly ("KIS", 1, 28)]